bearing-fault-detection/requirements.txt
torch>=1.7.0
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.5.0
scikit-learn>=0.23.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
websockets>=10.0
pywt>=1.1.0
tqdm>=4.50.0
requests>=2.25.0
paho-mqtt>=1.5.0
python-multipart>=0.0.5